use bedtools bamtofastq instead of tophas,modify download.sh

+++ download.sh
@@ -8,26 +8,28 @@
 # tar xvfz *.tar.gz
 # cd shenglinmei*
 # python setup.py install
+# bam2fastx not yet
 
 help()
 {
     cat <<HELP
 Instruction: download all the depend and install automatically
-              need root gcc numpy ghostscript and django
-USAGE EXAMPLE: bash install.sh -p pathd
+              need root gcc
+	      note: mdseqpos needs numpy ghostscript and django
+USAGE EXAMPLE: sudo bash install.sh -p pathd
     -p pathd: where to store download data
 HELP
-exit 0
+    exit 0
 }
 
 while [ -n "$1" ]; do
     case $1 in
-		-h)help;shift 1;;
-		-p)pathd=$2; shift 2;;
-		--)shift; break;; # options end
-		-*) echo "error: no such option $1. -h for help";exit 1;;
-		*) break;;
-		esac
+	-h)help;shift 1;;
+	-p)pathd=$2; shift 2;;
+	--)shift; break;; # options end
+	-*) echo "error: no such option $1. -h for help";exit 1;;
+	*) break;;
+    esac
 done
 
 if [ -z $pathd ]; then
@@ -35,30 +37,33 @@ if [ -z $pathd ]; then
     echo "please input the path for containing download data"
 fi
 
-# choose for all users or single user
-PS3="Choose (1-2)"
-echo "Choose from the list below."
-select name in global personal
-do
-    break
-done
-if [ $name = global ]; then
-    u=`whoami`
-    if [ $u = root ]; then
-        echo "install for all users"
-    else
-        echo "need root authority, ask your administrator!"
-        exit 1
-    fi
+## choose for all users or single user
+#PS3="Choose (1-2)"
+#echo "Choose from the list below."
+#select name in global personal
+#do
+#    break
+#done
+#if [ $name = global ]; then
+u=`whoami`
+if [ $u = root ]; then
+    echo "install for all users"
 else
-    echo "get ChiLin data to $pathd for single user"
+    echo "need root authority, ask your administrator!"
+    exit 1
 fi
+#else
+#    echo "get ChiLin data to $pathd for single user"
+#fi
 
 # Users could customize executive bin path
 read -p "choose where the bin is " bin
 if [ ! -n $bin ];then
-    echo "use default bin path /usr/bin"
-    bin=/usr/bin
+    if (echo $PATH|grep $bin) ; then echo 1;
+    else
+        echo "input bin not in $PATH, use default /usr/bin"
+        bin=/usr/bin
+    fi
 fi
 
 cp chilin/lib/db/*.bed $bin
@@ -77,27 +82,25 @@ cd $pathd
 # for judging machine type for download the right
 # source or binary
 if uname -a | grep Darwin; then
-    if uname -a | grep x86; then
-        machine="macOSX.ppc"
-	echo "Your machine type is 64 mac machine"
-    else
-        machine="macOSX.i386"
-        echo "Your machine type is 32 mac machine"
-    fi
+#    if uname -a | grep x86; then
+#        machine="macOSX.ppc"
+##	echo "Your machine type is 64 mac machine"
+#    else
+    machine="macOSX.i386"
+#        echo "Your machine type is 32 mac machine"
+    echo "your machine is mac"
+#    fi
 elif uname -a |grep Linux; then
     if uname -a | grep x86; then
         machine="linux.x86_64"
 	echo "Your machine type is linux machine"
     else
-        machine="linux.i386"
+        machine="Your machine type is linux.i386"
     fi
 fi
 
 get() {
     file=$1
-    # cmd="type -a $2"
-    # if ! $cmd > /dev/null
-    # then
     if ! wget --version >/dev/null 2>/dev/null ; then
 	if ! curl --version >/dev/null 2>/dev/null ; then
 	    echo "Please install wget or curl somewhere in your PATH"
@@ -109,30 +112,34 @@ get() {
 	wget $1 -O `basename $1`
 	return $?
     fi
-    # else
-    #     echo "$2 already installed"
-    # fi
 }
 
 install() {
     suffix=$1
     if [ ${suffix##*.} = py ];then
-	python setup.py install
+	echo "installining $2"
+	python setup.py install --prefix=`dirname $bin`
 	return $?
     elif [ ${suffix#*.} = cc ]; then
-	make
         name=`echo $2|cut -d . -f 1`
-	cp $name $bin
+	cd $name-*
+	make
+	if [ -d bin ]; then
+	    cp -r bin/* $bin
+	else
+	    cp $name $bin
+	fi
+	cd -
 	return $?
     else
 	cp `basename $2` $bin
 	return $?
     fi
-    cd -
 }
 
 # depend source code and binary
 # machine type specific 
+# mac_ppc version test failure, use i386
 URL_BASE="\
 http://bedtools.googlecode.com/files/BEDTools.v2.16.2.tar.gz \
 http://www.bioinformatics.babraham.ac.uk/projects/fastqc/fastqc_v0.10.1.zip \
@@ -140,24 +147,34 @@ http://hgdownload.cse.ucsc.edu/admin/exe/$machine/bedClip \
 http://hgdownload.cse.ucsc.edu/admin/exe/$machine/bedGraphToBigWig \
 "
 for i in $URL_BASE;do
-    if get $i; then
-        echo "download $i ready"
+    if [ ! -f `basename $i` ]; then
+        if get $i; then
+            echo "download $i ready"
+        fi
     fi
 done
+
 install other bedClip
 install other bedGraphToBigWig
-unzip fastqc_v0.10.1.zip -d $bin
+unzip fastqc_v0.10.1.zip -d $bin >/dev/null
 cd $bin
-chmod 755 fastqc
+chmod 755 $bin/FastQC/fastqc
 cd -
 
 ## for sourceforge special
-wget -c http://sourceforge.net/projects/bowtie-bio/files/bowtie/0.12.8/bowtie-0.12.8-src.zip/download -O bowtie.zip
-wget -c http://sourceforge.net/projects/samtools/files/samtools/0.1.18/samtools-0.1.18.tar.bz2/download -O samtools.tar.bz2
-unzip bowtie.zip
+if [ ! -f bowtie.zip ]; then
+    wget -c http://sourceforge.net/projects/bowtie-bio/files/bowtie/0.12.8/bowtie-0.12.8-src.zip/download -O bowtie.zip
+    if [ ! -f samtools.tar.bz2 ]; then
+	wget -c http://sourceforge.net/projects/samtools/files/samtools/0.1.18/samtools-0.1.18.tar.bz2/download -O samtools.tar.bz2
+    fi
+fi
+
+unzip bowtie.zip >/dev/null
+echo "install bowtie"
 install cc bowtie.zip
-tar xvfj samtools.tar.bz2
+tar xvfj samtools.tar.bz2 >/dev/null 2>&1
 install cc samtools.tar.bz2
+echo "install samtools"
 
 # bowtie index
 # you may add more index
@@ -168,21 +185,24 @@ hg19 \
 BowtieIndexBase="ftp://ftp.cbcb.umd.edu/pub/data/bowtie_indexes/"
 for index in $species; do
     if [ ! -f ${index}.ebwt.zip ]; then
-		if get ${BowtieIndexBase}/${index}.ebwt.zip; then
-			    echo "Bowtie index download successfully"
-					unzip ${BowtieIndexBase}/${index}.ebwt.zip -d $bin
-			fi
+	if get ${BowtieIndexBase}/${index}.ebwt.zip; then
+	    echo "Bowtie index download successfully"
+	    unzip ${BowtieIndexBase}/${index}.ebwt.zip -d $bin
+	fi
     fi
 done
 
 # github certificate macs2
 if type -a macs2
 then
-   echo "macs2 already install"
-else:
-   wget --no-check-certificate https://github.com/taoliu/MACS/zipball/master -O macs2.tar.gz
-   echo "Latest macs2 downloaded from github"
-   install py macs2.tar.gz
+    echo "macs2 already install"
+else
+    wget --no-check-certificate https://github.com/taoliu/MACS/zipball/master -O taoliu-MACS2.tar.gz
+    echo "Latest macs2 downloaded from github"
+    export CC=gcc  # upper case
+    tar xvfz taoliu-MACS2.tar.gz>/dev/null
+    cd taoliu*
+    install py macs2
 fi
 
 # mercurial cistrome-applications
@@ -194,20 +214,23 @@ fi
 if [ ! -f cistrome-app.tar.gz ]; then
     wget --no-check-certificate \
         https://bitbucket.org/cistrome/cistrome-applications-harvard/get/e82ed15a486b.tar.gz -O cistrome-app.tar.gz
-    tar xvfz cistrome-app.tar.gz
-    cd cistrome-applications-harvard*
-    # ceas
-    cd published-packages/CEAS
-    python setup.py install --prefix=`dirname $bin`
-    cd gdb && gunzip hg19* && gunzip mm9*
-    cp hg19* mm9* $bin
-    cd ../../cistrome-extra-apps
-    python setup.py install --prefix=`dirname $bin`
-    # mdseqpos problem
-    export CC=gcc  # upper case
-    cd ../../mdseqpos
-    python setup.py install --prefix=`dirname $bin`
 fi
+tar xvfz cistrome-app.tar.gz >/dev/null 2>&1
+cd cistrome-cistrome-applications-harvard*
+# ceas
+cd published-packages/CEAS
+install py CEAS
+cd gdb && gunzip *.refGene.gz
+cp -rf *.refGene $bin
+cd ../../../cistrome-extra-apps built-ins
+install py built-ins
+
+# mdseqpos problem: gcc setting and settings.py
+# 
+export CC=gcc  # upper case
+cd ../mdseqpos
+cp lib/settings.py.example lib/settings.py
+install py mdseqpos
 #
 # special download need copy sub directory
 # conservation Phascon bigwiggle
@@ -238,15 +261,14 @@ chrX \
 chrY \
 chrM"
 for s in $species;do
-	for c in $BASE_CHRS;do
-		mkdir conservation_$s 2>/dev/null
-		# vertebrate
-		wget -c http://cistrome.org/~hanfei/conservation/$s/vertebrate/${c}.bw -O conservation_$s/${c}.bw
+    for c in $BASE_CHRS;do
+	mkdir conservation_$s 2>/dev/null
+	wget -c http://cistrome.org/~hanfei/conservation/$s/vertebrate/${c}.bw -O conservation_$s/${c}.bw
     done
     cp -r conservation_$s $bin
 done
 
-cd - #back to the chilin directory
+cd ../../.. #back to the chilin root directory
 
 echo "\
 [UserInfo]
@@ -259,9 +281,11 @@ controlpath =
 OutputDirectory = 
 # should add fastq or bam suffix
 [bowtie]
-BAMTOFASTQ = $bin/bam2fastx
+#BAMTOFASTQ = $bin/bam2fastx
+# latest bedtools
+BAMTOFASTQ = $bin/bamToFastq
 BOWTIE_MAIN = $bin/bowtie
-BOWTIE_GENOME_INDEX_PATH = $bin/BowtieIndex/{{ species }}
+BOWTIE_GENOME_INDEX_PATH = $bin/{{ species }}
 nBOWTIE_MAX_ALIGNMENT = 1
 
 [samtools]
@@ -296,7 +320,7 @@ CONSERV_PLOT_PHAST_PATH = $bin/conservation_{{ species }}
 # use vertebrate Phascon
 
 [correlation]
-WIG_CORRELATION_MAIN = $bin/wig_correlation_bigwig_input_only.py
+WIG_CORRELATION_MAIN = $bin/bigwig_correlation.py
 WIG_CORRELATION_STEP = 10
 WIG_CORRELATION_METHOD = mean 
 WIG_CORRELATION_MIN = 2
@@ -319,10 +343,14 @@ SEQPOS_PVALUE_CUTOFF = 0.001
 SEQPOS_MOTIF_DB_SELECTION = cistrome.xml
 
 [QC]
-FASTQC_MAIN = $bin/fastqc
+FASTQC_MAIN = $bin/FastQC/fastqc
 FILTERDUP_SPECIES = {{ filterdup_species }}
-" > ChiLin.conf.new
-if $?
-then
-   echo "ChiLin data config succeed!:)"
+" > chilin/lib/template/ChiLinjinja.conf.new
+
+echo "========================================="
+#python setup.py install # ChiLin
+
+if [ $? -eq 0 ]; then
+    echo "ChiLin data config succeed!:)"
 fi
+echo "Please run ChiLin setup.py"
